Added functionality if solution not satisfactory
If the user who posted the bug tries the fix and isn't satisfied with the solution, they need to be able to tell the solver/ other users why they aren't satisfied.
I want them to be able to post an update like the solver does but the update class takes a solver object and the poster is not a solver.
will have to change models to reflect this so commiting this pre-model change